Extract BookView
