+++ app/src/main/java/com/github/geekarist/henripotier/Notifiable.java
@@ -1,0 +1,5 @@
+package com.github.geekarist.henripotier;
+
+public interface Notifiable {
+    void notifyChange();
+}
