+++ app/src/main/res/layout/activity_cart_item.xml
@@ -5,7 +5,7 @@
     android:orientation="horizontal">
 
     <ImageView
-        android:id="@+id/cartItemImageView"
+        android:id="@+id/cart_item_image_view"
         android:layout_width="90dp"
         android:layout_height="fill_parent"
         android:layout_marginBottom="5dp"
@@ -18,14 +18,14 @@
         android:orientation="vertical">
 
         <TextView
-            android:id="@+id/cartItemTitleView"
+            android:id="@+id/cart_item_title_view"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="Large Text"
             android:textAppearance="?android:attr/textAppearanceLarge" />
 
         <TextView
-            android:id="@+id/cartItemPriceView"
+            android:id="@+id/cart_item_price_view"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:text="Medium Text"
